+++ wtfix/core/exceptions.py
@@ -35,22 +35,21 @@ class ValidationError(Exception):
 
 
 class MessageProcessingError(Exception):
+    """
+    Used to signal an abnormal interrupt while a message is being processed. This exception will not cause the
+    pipeline to be stopped and the default implementation will just log the error before proceeding to the next
+    message.
+    """
     pass
 
 
 class StopMessageProcessing(Exception):
-    pass
+    """
+    Used to stop a message from propagating further up or down the pipeline.
 
-
-class InvalidMessage(Exception):
-    pass
-
-
-class SignalStop(Exception):
-    pass
-
-
-class WaitTimeout(BaseException):
+    This should be used to interrupt message processing during normal operation (i.e. as part of an optimization
+    or if the message is not relevant to other applications in the pipeline.
+    """
     pass
 
 
@@ -60,9 +59,9 @@ class InvalidField(Exception):
 
 class SessionError(Exception):
     """
-    Fatal session error from which no recovery is possible
+    Fatal session error from which no recovery is possible. A SessionError exception won't be handled by the
+    pipeline and will cause an abnormal termination.
     """
-
     pass